+++ app/Models/CreditLimitsValidation.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class CreditLimitsValidation extends Model
+{
+    use HasFactory;
+
+    public $fillable = ['credit_limit', 'user_id'];
+
+    public function user()
+    {
+        return $this->belongsTo(User::class, 'user_id');
+    }
+}
